# Faster ONNX inference with bindings