--- a/ultralytics/nn/autobackend.py
+++ b/ultralytics/nn/autobackend.py
@@ -193,6 +193,19 @@ class AutoBackend(nn.Module):
             session = onnxruntime.InferenceSession(w, providers=providers)
             output_names = [x.name for x in session.get_outputs()]
             metadata = session.get_modelmeta().custom_metadata_map
+            io = session.io_binding()
+            bindings = []
+            for output in session.get_outputs():
+                y_tensor = torch.empty(output.shape, dtype=torch.float16 if fp16 else torch.float32).to(device).contiguous()
+                io.bind_output(
+                    name=output.name,
+                    device_type=device.type,
+                    device_id=device.index if cuda else 0,
+                    element_type=np.float16 if fp16 else np.float32,
+                    shape=tuple(y_tensor.shape),
+                    buffer_ptr=y_tensor.data_ptr(),
+                )
+                bindings.append(y_tensor)
 
         # OpenVINO
         elif xml:
@@ -476,8 +489,20 @@ class AutoBackend(nn.Module):
 
         # ONNX Runtime
         elif self.onnx:
-            im = im.cpu().numpy()  # torch to numpy
-            y = self.session.run(self.output_names, {self.session.get_inputs()[0].name: im})
+            device_type=im.device.type
+            device_id=im.device.index if device_type == "cuda" else 0
+    
+            self.io.bind_input(
+                name='images',
+                device_type=device_type,
+                device_id=device_id,
+                element_type=np.float16 if self.fp16 else np.float32,
+                shape=tuple(im.shape),
+                buffer_ptr=im.data_ptr(),
+                )
+
+            y = self.bindings
+            self.session.run_with_iobinding(self.io)
 
         # OpenVINO
         elif self.xml: