Closes https://github.com/ultralytics/ultralytics/issues/17127

Uses bindings to reduce CPU-GPU transfer overheads.

**Export**
```python
from ultralytics import YOLO

model = YOLO("yolo11n.pt")
model.export(format="onnx")
```

**Before changes**

```python
from ultralytics import YOLO

model = YOLO("yolo11n.onnx", task="detect")
results = model.val(data="coco.yaml", device=0)
```
```
Speed: 0.3ms preprocess, 9.0ms inference, 0.0ms loss, 1.6ms postprocess per image
```

**After**
```python
from ultralytics import YOLO

model = YOLO("yolo11n.onnx", task="detect")
results = model.val(data="coco.yaml", device=0)
```
```
Speed: 0.3ms preprocess, 5.9ms inference, 0.0ms loss, 1.6ms postprocess per image
```

<!--
Thank you 🙏 for your contribution to [Ultralytics](https://www.ultralytics.com/) 🚀! Your effort in enhancing our repositories is greatly appreciated. To streamline the process and assist us in integrating your Pull Request (PR) effectively, please follow these steps:

1. Check for Existing Contributions: Before submitting, kindly explore existing PRs to ensure your contribution is unique and complementary.
2. Link Related Issues: If your PR addresses an open issue, please link it in your submission. This helps us better understand the context and impact of your contribution.
3. Elaborate Your Changes: Clearly articulate the purpose of your PR. Whether it's a bug fix or a new feature, a detailed description aids in a smoother integration process.
4. Ultralytics Contributor License Agreement (CLA): To uphold the quality and integrity of our project, we require all contributors to sign the CLA. Please confirm your agreement by commenting below:

    I have read the CLA Document and I sign the CLA

For more detailed guidance and best practices on contributing, refer to our ✅ [Contributing Guide](https://docs.ultralytics.com/help/contributing/). Your adherence to these guidelines ensures a faster and more effective review process.
--->


## 🛠️ PR Summary

<sub>Made with ❤️ by [Ultralytics Actions](https://github.com/ultralytics/actions)<sub>

### 🌟 Summary
Enhanced ONNX Runtime support with IO binding in the autobackend module.

### 📊 Key Changes
- Implemented IO binding for ONNX models, allowing direct tensor manipulation on the device.
- Replaced numpy conversion with `bind_input` method for ONNX inference, improving execution efficiency.
- Added support for tensor allocation and management directly with `onnxruntime.InferenceSession`.

### 🎯 Purpose & Impact
- **Faster Inference**: By eliminating unnecessary data transfers and conversions, this update accelerates model inference time, particularly on GPU.
- **Optimized Memory Usage**: Direct device tensor binding can lead to more efficient memory usage, benefiting both performance and resource management.
- **Improved Scalability**: These changes make the model backend more robust and scalable, accommodating larger model inputs with better performance.