@UltralyticsAssistant 2024-10-26
👋 Hello @Y-T-G, thank you for submitting an `ultralytics/ultralytics` 🚀 PR! This is an automated response to help guide the review process, and an Ultralytics engineer will assist you soon.

To ensure a smooth integration of your contributions, please review the checklist below:

- ✅ **Define a Purpose**: Clearly articulate the purpose of your fix or enhancement in your PR description, and link to any relevant issues like [this one](https://github.com/ultralytics/ultralytics/issues/17127). Ensure your commit messages are precise and align with the project's standards.
- ✅ **Synchronize with Source**: Make sure your PR is up-to-date with the `main` branch of `ultralytics/ultralytics`. If there are discrepancies, you can resolve them by clicking the 'Update branch' button or by executing `git pull` and `git merge main` locally.
- ✅ **Ensure CI Checks Pass**: Verify all Ultralytics [Continuous Integration (CI)](https://docs.ultralytics.com/help/CI/) checks are successful. If any checks fail, please identify and fix the issues.
- ✅ **Update Documentation**: Update the [documentation](https://docs.ultralytics.com) as needed for any new or updated features.
- ✅ **Add Tests**: Include or modify tests to ensure your changes are adequately covered, and validate that all tests pass.
- ✅ **Sign the CLA**: If it's your inaugural PR with Ultralytics, please sign the [Contributor License Agreement](https://docs.ultralytics.com/help/CLA/) by stating, "I have read the CLA Document and I sign the CLA" in a new comment.
- ✅ **Minimize Changes**: Modify only the code essential for your bug fix or feature enhancement. 🧘‍♂️ As Bruce Lee said, _"It is not daily increase but daily decrease, hack away the unessential."_

For comprehensive guidance, take a look at our [Contributing Guide](https://docs.ultralytics.com/help/contributing). If you have any queries, don't hesitate to ask. Thank you for contributing to Ultralytics! 🚀✨