@Y-T-G 2024-10-28
@Burhan-Q Interesting. I'm not able to get a consistent time on `coco128.yaml` on CPU. Sometimes the postprocessing is faster in the `main` and sometimes with this PR.

One of the runs:
`main`
`Speed: 1.1ms preprocess, 65.1ms inference, 0.0ms loss, 4.5ms postprocess per image`

`PR`
`Speed: 1.2ms preprocess, 61.1ms inference, 0.0ms loss, 3.0ms postprocess per image`

Particularly, if you load the model in Python and run val the second time, it's much better than the first time.