@Y-T-G 2024-10-28
@Laughing-q Hmm. It seems like your ONNX InferenceSession failed to use CUDAExexutionProvider.

https://github.com/ultralytics/ultralytics/blob/e1e54602dfbf6b9c52aaaa340681a5a8b8f089f4/ultralytics/nn/autobackend.py#L192-L193

Do you have onnxruntime-gpu?

Regardless, I should add a check and probably a warning too for this scenario.