@Laughing-q 2024-10-28
@Y-T-G ok I reinstalled the onnxruntime package and it works! I do get faster speed than main branch.