@glenn-jocher 2024-10-29
@Y-T-G @Laughing-q maybe there's a better way to automatically handle the execution providers. I've updated it to now retrieve all available providers. On macOS for example there are additional ones here I didn't see before:
```
['CoreMLExecutionProvider', 'AzureExecutionProvider', 'CPUExecutionProvider']
```