@glenn-jocher 2024-10-29
@Y-T-G @Laughing-q wow,  just independently allowing for a wider set of execution providers (just my last commit, not the rest of the PR) has this change (also slower pre and post processing but much faster inference)

<img width="1510" alt="Screenshot 2024-10-29 at 13 19 28" src="https://github.com/user-attachments/assets/32b148b9-1759-40a7-a062-5f5bd5c7897f">