@Y-T-G 2024-10-29
@glenn-jocher I guess it's using CoreMLProvider in your case which should be faster than CPU