@YOOHYOJEONG 2024-10-31
@Y-T-G 

Hi. 
I checked that the code is currently modified to this PR.
I tried benchmark again after pull with this version code, and I got the following error.


Loading /ultralytics/weights/coco_pretrained/yolov11/yolo11n.onnx for ONNX Runtime inference...
Preferring ONNX Runtime TensorrtExecutionProvider
2024-10-31 00:05:37.806269014 [E:onnxruntime:Default, provider_bridge_ort.cc:1978 TryGetProviderInfo_TensorRT] /onnxruntime_src/onnxruntime/core/session/provider_bridge_ort.cc:1637 onnxruntime::Provider& onnxruntime::ProviderLibrary::Get() [ONNXRuntimeError] : 1 : FAIL : Failed to load library libonnxruntime_providers_tensorrt.so with error: libnvinfer.so.10: cannot open shared object file: No such file or directory

*************** EP Error ***************
EP Error /onnxruntime_src/onnxruntime/python/onnxruntime_pybind_state.cc:490 void onnxruntime::python::RegisterTensorRTPluginsAsCustomOps(PySessionOptions&, const onnxruntime::ProviderOptions&) Please install TensorRT libraries as mentioned in the GPU requirements page, make sure they're in the PATH or LD_LIBRARY_PATH, and that your GPU is supported.
 when using ['TensorrtExecutionProvider', 'CUDAExecutionProvider', 'CPUExecutionProvider']
Falling back to ['CUDAExecutionProvider', 'CPUExecutionProvider'] and retrying.
****************************************   

May I know your onnx version and tensorRT version?


But even if this error occurs, it doesn't mean that it can't be inferred as onnx.
It appears to be deduced using GPU, much faster than the traditional main.

But it's still slower than pytorch..   
![image](https://github.com/user-attachments/assets/eda5c248-2ff9-4695-9fb1-1593acef4bdf)

berfoe : 
https://github.com/ultralytics/ultralytics/issues/17127