@Y-T-G 2024-10-31
@YOOHYOJEONG The PyTorch model also runs with minimum rectangle, so it doesn't use the full 640x640 unlike ONNX which can give it an edge. You can try setting `rect=False` to `model.val` for PyTorch.