@YOOHYOJEONG 2024-11-08
@Y-T-G 
Hi. 

`rect=False` doesn't seem to progress to 640X640 even with this.
The actual output value is 384x640.

validate code   
'''python
from ultralytics import YOLO

# Load a model
model = YOLO("/ultralytics/DATA/weights/coco_pretrained/yolov11/yolo11n.pt")

# Customize validation settings
val_results = model.val(data="/ultralytics/coco.yaml", imgsz=640, rect=False, device="0")
'''
   
And if you look at the argument document(https://docs.ultralytics.com/modes/val/#arguments-for-yolo-model-validation), the default value is False, right?