@Y-T-G 2024-11-08
@YOOHYOJEONG You also probably need to change the `Letterbox` function to not use `auto=True` by default.